(fix): remove runtime `pytest` dep
