+++ src/testing/anndata/_doctest.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from collections.abc import Callable
+from typing import TypeVar
+
+F = TypeVar("F", bound=Callable)
+
+
+def doctest_needs(mod: str) -> Callable[[F], F]:
+    """Mark function with doctest dependency."""
+
+    def decorator(func: F) -> F:
+        func._doctest_needs = mod
+        return func
+
+    return decorator
